+++ src/app/js/root-admin/types.ts
@@ -1,0 +1,12 @@
+export type Tenant = {
+    _id: string;
+    name?: string;
+    description: string;
+    author: string;
+    username: string;
+    password: string;
+    createdAt?: string;
+    dataset?: number;
+    published?: boolean;
+    totalSize?: number;
+};
